Add raw pairing and connection flow example
This is a traffic capture of the connection process (including
pairing) between the Remote app and tvOS simulator. It's raw with
fully decoded protobuf message and is meant to serve as a way to know
which messages are passed back and forth when a connection is made.
More example can be added later (maybe also move them to separate
files).